Added !bacon command!
This reads the links in BaconLinks.json, and picks a random one to post to the server. The user may also limit the search by adding an optional tag like 'skyrim' to only get skyrim related videos!
To achieve this the kotlinx.serialization dependency was added to maven. This should make it muche easier to read .json files in the future :)