Constructed a parser for bulbapedia source (that almost works)!
The parser translates page source from bulbapedia stored in raw/ListOfPokemon.txt to PokemonText.json that is indexed on generation and type(s). The format is readable, although has a dangling comma on the end of closing expressions, but this should not be hard to solve. This parser may later be used when getting random pokemon.
The JSON file may be updated by using the command "!dev pp", and should be scalable for more pokemon. A limit on how many generations that should be fetched is also included in "PokemonFixedParser" as "lastGen"