Cleaned up waypoint nodes so that they are created from varargs instead of list
As the list only is used to pass arguments, this should be a cleaner and faster way to create waypoint nodes. This is not applied to regular nodes as these may have an optional bullet point and can therefore not use varargs. (I might hard code the bullet point in later so that varargs are possible here as well)